+++ multiworld/envs/assets/sawyer_torque/sawyer_reach_torque.xml
