+++ multiworld/envs/assets/sawyer_torque/sawyer_reach_torque_with_gripper.xml
