+++ multiworld/envs/mujoco/__init__.py
@@ -61,7 +61,7 @@ def register_custom_envs():
 
     register(
         id='SawyerReachTorqueEnv-v0',
-        entry_point='multiworld.envs.mujoco.dynamic_robotics.sawyer_reach_torque_env:SawyerReachTorqueEnv',
+        entry_point='multiworld.envs.mujoco.sawyer_torque.sawyer_reach_torque_env:SawyerReachTorqueEnv',
         tags={
             'git-commit-hash': 'dd10a1f',
             'author': 'murtaza'
@@ -74,7 +74,7 @@ def register_custom_envs():
 
     register(
         id='SawyerReachTorqueGripperEnv-v0',
-        entry_point='multiworld.envs.mujoco.dynamic_robotics.sawyer_torque_reacher_with_gripper:SawyerReachTorqueGripperEnv',
+        entry_point='multiworld.envs.mujoco.sawyer_torque.sawyer_torque_reacher_with_gripper:SawyerReachTorqueGripperEnv',
         tags={
             'git-commit-hash': 'dd10a1f',
             'author': 'murtaza'
